pyproject.toml
[project]
name = "dta-backend"
version = "0.1.0"
description = "DTA Backend Server"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "bcrypt>=5.0.0",
    "fastapi[standard]>=0.121.1",
    "pydantic-settings>=2.12.0",
    "python-jose>=3.5.0",
    "sqlalchemy>=2.0.44",
]
